update HikingPoint property price from Double to Long

+++ src/main/java/com/abdav/giri_guide/model/response/HikingPointResponse.java
@@ -5,7 +5,7 @@ public record HikingPointResponse(
         String mountainId,
         String name,
         String coordinate,
-        Double price
+        Long price
 
 ) {
 }
